01-Login/requirements.txt
django==1.11.8
social-auth-app-django
python-jose
six
python-dotenv
